Work with HTML instead of stripped text for char dir test
Sadly, xpath seems to not like utf-8 much, so non ascii characters
are not parsed correctly in some servers (though surprisingly not
in my local machine, for one) -- which makes the test invalid, as
all characters are read as aschii LTR.

We are going to have to use the raw HTML for now, even though this
tends to produce a timeout if the page is really long.

I'm researching possible solutions to this.